+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Core: fix searching for partial email addresses in global search
+- Core: fix boolean select filters in custom filters
 
 11-03-2024: 6.7.73
 - Core: Make add button more visible in link browser